Re-namespace and categorize Combat Events to keep things organized; also new test event
