+++ Assets/Scripts/Combat/CombatController.cs
@@ -6,6 +6,8 @@ using System.Collections.Generic;
 using UnitGroups;
 // NOTE(jordan): List.All
 using System.Linq;
+using EventSystem;
+using Combat.Events;
 
 /// <summary>
 /// Controller for the Combat Scene to figure out whose turn it is, etc.
@@ -93,6 +95,7 @@ public class CombatController : MonoBehaviour {
         InitializeUnitGroups();
 
         CombatTimer.Every(turnInterval, OnTurn);
+        EventManager.On<TurnOver>(() => Debug.Log(string.Format("Turn swap! {0}", TurnOverMessage())));
     }
 
     /// <summary>
@@ -125,6 +128,8 @@ public class CombatController : MonoBehaviour {
         UIController.SetMessage(TurnOverMessage());
         // NOTE(jordan): show message for 1/2 turnInterval seconds (needs to be less than turnInterval or it will show and hide at the same time ;) )
         UIController.Show(turnInterval / 2f);
+
+        EventManager.Trigger<TurnOver>();
     }
 
     /// <summary>
