+++ Assets/Scripts/EventSystem/EventSystem.cs
@@ -2,26 +2,26 @@
 using System;
 using System.Collections.Generic;
 
-namespace Events {
-    public class EventSystem : MonoBehaviour {
-        public List<IEventListener> Listeners = new List<IEventListener>();
+namespace EventSystem {
+    public static class EventManager {
+        public static List<IEventListener> Listeners = new List<IEventListener>();
 
-        private IEventListener GetListenerFor<T> () where T: IEvent {
+        private static IEventListener GetListenerFor<T> () where T: IEvent {
             Type eventType = typeof(T);
             return Listeners.Find((listener) => listener.EventType == eventType);
         }
 
-        public void On<T> (Action callback) where T: IEvent {
+        public static void On<T> (Action callback) where T: IEvent {
             IEventListener l = GetListenerFor<T>();
             l.Callbacks.Add(callback);
         }
 
-        public bool Off<T> (Action callback) where T: IEvent {
+        public static bool Off<T> (Action callback) where T: IEvent {
             IEventListener l = GetListenerFor<T>();
             return l.Callbacks.Remove(callback);
         }
 
-        public void Trigger<T> () where T: IEvent {
+        public static void Trigger<T> () where T: IEvent {
             IEventListener l = GetListenerFor<T>();
             l.Invoke();
         }
