+++ Assets/Scripts/EventSystem/IEvent.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace Events {
+namespace EventSystem {
     public interface IEvent {
         IEventListener Listener { get; }
     }
